factored out main into own file
